requirements.txt
amqp==5.0.9
ansible==4.10.0
ansible-base==2.10.17
ansible-core==2.11.12
ansible-runner==1.4.6
asgiref==3.4.1
backports.zoneinfo==0.2.1
bcrypt==3.2.0
billiard==3.6.4.0
brew==0.1.3
cached-property==1.5.2
celery==5.0.2
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
Deprecated==1.2.13
Django==3.2.14
django-appconf==1.0.4
django-auth-ldap==4.0.0
django-extensions==3.1.5
django-filter==21.1
django-forms-dynamic==1.0.0
django-materializecss-form==1.1.17
django-mssql-backend==2.8.1
django-python3-ldap==0.12.0
django-redis==5.2.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-select2==7.10.0
django-sslserver==0.22
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
docutils==0.16
ftfy==5.8
graphviz==0.17
idna==2.10
importlib-metadata==4.6.1
importlib-resources==3.3.0
itypes==1.2.0
Jinja2==3.0.3
kombu==5.1.0
lockfile==0.12.2
mailchecker==3.3.15
Markdown==3.3.6
MarkupSafe==2.0.1
mssql-django==1.1.2
netaddr==0.8.0
openapi-codec==1.3.2
packaging==21.3
pan-os-python==1.7.3
pan-python==0.17.0
pandevice==0.14.0
paramiko==2.7.2
pexpect==4.8.0
phonenumbers==8.12.11
Pillow==8.3.2
prettytable==1.0.1
prompt-toolkit==3.0.19
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
pyeapi==0.8.4
PyNaCl==1.4.0
pyodbc==4.0.32
pyparsing==3.0.9
pypyodbc==1.3.2
python-benedict==0.22.0
python-daemon==2.2.4
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2021.3
PyYAML==6.0
redis==3.5.3
requests==2.24.0
resolvelib==0.5.4
simplejson==3.17.6
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.1
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
xmltodict==0.12.0
zipp==3.4.0
